requirements.txt
django
djangorestframework==3.11.0
ipdb==0.13.3
Pillow==7.1.1
django-cors-headers
drf-yasg
mutagen
psycopg2
